requirements.txt
Flask==3.0.2
Flask-CORS==5.0.0
Flask-Mail==0.9.1
flask-apscheduler==1.13.1
flask-socketio==5.3.6
python-socketio==5.9.0
python-engineio==4.8.0
eventlet==0.33.3
firebase-admin==6.4.0
google-cloud-storage==2.14.0
pymongo>=4.6.1,<5.0.0
openai==1.54.0
python-dotenv==1.0.1
twilio==9.0.0
gunicorn==21.2.0
pytz==2024.1
Werkzeug==3.0.1
pandas==2.0.3
openpyxl==3.1.5
numpy==1.24.3
dnspython==2.4.2
